Переход на локальное кэширование в фаербейз, уход от локального хранения в рум. Интерфейсы для функционала фаербейз. Рефакторинг

+++ domain/src/main/java/com/taxapprf/domain/report/ReportModel.kt
@@ -2,5 +2,5 @@ package com.taxapprf.domain.report
 
 data class ReportModel(
     val year: String,
-    val tax: Long,
+    val tax: Double,
 )
